% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_say.R
\name{animal_say}
\alias{animal_say}
\title{Call an animal and let it speak for you!}
\usage{
animal_say(number, string)
}
\arguments{
\item{number}{1 or 2 in our case to specify an animal}

\item{string}{a string stating what you want to say}
}
\description{
Call an animal and let it speak for you!
}
\examples{

animal_say(1, "hahaha!!!")
}
