% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_temp.R
\name{convert_temp}
\alias{convert_temp}
\title{Convert temperarure}
\usage{
convert_temp(temperature, method)
}
\arguments{
\item{temperature}{A number represents temperature.}

\item{method}{A string.}
}
\value{
The number of temp converted.
}
\description{
Convert temperarure
}
\examples{
convert_temp(1,"f2c")
}
